requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-multipart
python-dotenv
requests
twilio
motor
